added code to Copy translationHelps from Static to User Resources
