+++ src/js/actions/ActionTypes.js
@@ -28,7 +28,6 @@ const ActionTypes = {
   CLOSE_POPOVER: "CLOSE_POPOVER",
   SAVE_MODULE_VIEW: "SAVE_MODULE_VIEW",
   ADD_NEW_BIBLE_TO_RESOURCES: "ADD_NEW_BIBLE_TO_RESOURCES",
-  ADD_NEW_RESOURCE: "ADD_NEW_RESOURCE",
   SET_SAVE_PATH_LOCATION: "SET_SAVE_PATH_LOCATION",
   UPDATE_PROGRESS: "UPDATE_PROGRESS",
   DONE_MODULES: "DONE_MODULES",
